Seperate WebviewManager from Webview
